fix all costs from long to double becouse of decimals

+++ src/main/java/nl/autogarage/finalassignmentbackendmain/dto/inputDto/InspectionInputDto.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class InspectionInputDto {
 
     private Long id;
-    private Long costEstimate;
+    private Double costEstimate;
     private String inspectionDescription;
     private boolean inspectionApproved;
     private boolean isInspected;
