+++ src/main/java/nl/autogarage/finalassignmentbackendmain/dto/outputDto/RepairOutputDto.java
@@ -16,7 +16,7 @@ public class RepairOutputDto {
     private Long id;
     private String repairDescription;
 
-    private Long partRepairCost;
+    private Double partRepairCost;
     private boolean repairFinished;
